Add base files
These notes are a partial copy of a previous workshop I have run in the past.  These will be edited to fit the current lectures about advanced Git workflows.

There is a basic python package with a single function defined with a simple unit test and a doc string.

The docs are st up with sphinx to display the lecture notes and the module doc strings.

Hopefully all the github actions work as expected...